+++ src/oca/go/src/goca/virtualrouter.go
@@ -35,7 +35,7 @@ type VirtualRouterPool struct {
 
 // VirtualRouter represents an OpenNebula VirtualRouter
 type VirtualRouter struct {
-	ID          uint            `xml:"ID"`
+	ID          int             `xml:"ID"`
 	UID         int             `xml:"UID"`
 	GID         int             `xml:"GID"`
 	UName       string          `xml:"UNAME"`
@@ -62,13 +62,13 @@ func (c *Controller) VirtualRouters() *VirtualRoutersController {
 }
 
 // VirtualRouter returns a VirtualRouter controller.
-func (c *Controller) VirtualRouter(id uint) *VirtualRouterController {
+func (c *Controller) VirtualRouter(id int) *VirtualRouterController {
 	return &VirtualRouterController{c, id}
 }
 
 // VirtualRouterByName returns a VirtualRouter By name
-func (c *Controller) VirtualRouterByName(name string, args ...int) (uint, error) {
-	var id uint
+func (c *Controller) VirtualRouterByName(name string, args ...int) (int, error) {
+	var id int
 
 	vrouterPool, err := (&VirtualRoutersController{c}).Info(args...)
 	if err != nil {
@@ -143,7 +143,7 @@ func (vc *VirtualRouterController) Info() (*VirtualRouter, error) {
 // Create allocates a new virtual router. It returns the new Virtual Router ID
 // * name: name of the virtualrouter
 // * tpl: template of the virtualrouter
-func (vc *VirtualRoutersController) Create(name string, tpl *DynamicTemplate) (uint, error) {
+func (vc *VirtualRoutersController) Create(name string, tpl *DynamicTemplate) (int, error) {
 	if tpl == nil {
 		return 0, fmt.Errorf("VirtualRouter Create: nil template arg")
 	}
@@ -154,7 +154,7 @@ func (vc *VirtualRoutersController) Create(name string, tpl *DynamicTemplate) (u
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // Update replaces the cluster cluster contents.
@@ -198,7 +198,7 @@ func (vc *VirtualRouterController) Delete() error {
 // * name: Name for the VM instances. If it is an empty string OpenNebula will set a default name. Wildcard %i can be used.
 // * hold: False to create the VM on pending (default), True to create it on hold.
 // * extraTpl: A string containing an extra template to be merged with the one being instantiated. It can be empty.
-func (vc *VirtualRouterController) Instantiate(number, tplid int, name string, hold bool, extraTpl *VMTemplate) (uint, error) {
+func (vc *VirtualRouterController) Instantiate(number, tplid int, name string, hold bool, extraTpl *VMTemplate) (int, error) {
 	extraTplStr := ""
 	if extraTpl != nil {
 		extraTplStr = extraTpl.String()
@@ -209,7 +209,7 @@ func (vc *VirtualRouterController) Instantiate(number, tplid int, name string, h
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // AttachNic attaches a new network interface to the virtual router and the virtual machines.
@@ -220,7 +220,7 @@ func (vc *VirtualRouterController) AttachNic(n *NIC) error {
 
 // DetachNic detaches a network interface from the virtual router and the virtual machines
 // * nicid: NIC ID to detach
-func (vc *VirtualRouterController) DetachNic(nicid uint) error {
+func (vc *VirtualRouterController) DetachNic(nicid int) error {
 	_, err := vc.c.Client.Call("one.vrouter.detachnic", vc.ID, nicid)
 	return err
 }
